[chore]: moving tests to tests/ directory and introducing src/ (#22)
See title. No code changes
